api/wwcodesvtools/requirements.txt
Django==3.2.23
django-environ==0.11.2
djangorestframework==3.14.0
pre-commit==2.7.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.0
flake8==6.1.0
gunicorn==21.2.0
psycopg2==2.9.9
whitenoise==6.6.0